+++ projects/formkit/src/lib/components/form-field/form-field.component.spec.ts
@@ -25,24 +25,23 @@ const field: ISingleField<any, any, any> = {
       return undefined;
     }
   },
-  messages: [
+  messages: (payload) => ([
     {
-      show: ({control}) => (control.value !== 'initial-value'),
+      show: (payload.control.value !== 'initial-value'),
       type: FieldMessageType.Information,
-      text: (payload) => 'this is a information message'
+      text: 'this is a information message'
     },
     {
       show: true,
       type: FieldMessageType.Warning,
-      text: (payload) => 'this is a warning that must always show'
+      text: 'this is a warning that must always show'
     }
-  ]
+  ])
 };
 
 describe('FieldComponent', () => {
   let component: FormFieldComponent;
   let fixture: ComponentFixture<FormFieldComponent>;
-  let events$: Subject<FormEvent>;
   let control: FormControl;
   let service: FormService;