+++ projects/showcase/src/app/components/inputs/e2e.required.inputs.ts
@@ -10,7 +10,9 @@ export const requiredFormFields: FormFields<RequiredForm> = {
   input: {
     type: FieldType.Text,
     validators: [Validators.email],
-    required: values => values.checkbox
+    status: ({ values }) => ({
+      required: values.checkbox
+    })
   },
   checkbox: {
     type: FieldType.Checkbox,